feat: adicionando logica de entrada na masmorra
